[chore#159221189]Configred pytesting
