+++ .coverage
@@ -1,0 +1,1 @@
+!coverage.py: This is a private format, don't read it directly!{"lines":{"I:\\andela_bootcamp\\My_Diary\\API\\tests\\test_app.py":[1,2,4,9,17,22,27,32,37,42,47,6,7,13,15,19,20,24,25,29,30,34,35,39,40,44,45,50]}}
